Title: 합과 차의 나눗셈, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240107/합과 차의 나눗셈/divide-of-sum-and-sub.cpp
@@ -8,7 +8,7 @@ int main() {
     cout.fixed;
     cout << precision(2);
 
-    cout << (a+b) / (a-b);
+    cout << (double)(a+b) / (a-b);
     // 여기에 코드를 작성해주세요.
     return 0;
 }
